Enforce unbuffered L2CAP socket for GATT
The GATT protocol does not enforce a length indicator on each GATT
command. Usually this is not a problem because each request is followed
by a response and therefore we can assume the entire packet content is
a single response/request.

However the GATT WRITE_COMMAND does not have a response. Multiple
requests can be sent in quick follow-up. If buffering is utilized the
consecutive commands may be merged and received as single data blob on
the remote device. Since Write commands don't have a length indicator
they cannot be distinguished and ultimately the write command will set
the wrong value.

Change-Id: I02f3cd3cfaedfeae6a40272f5d22d3d88c39aa55
Task-number: QTBUG-49650
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>