BLE: Write the whole 31 bytes of advertising data to the controller.
The specification says: "Only the significant part of the Advertising
Data is transmitted in the advertising packets". This seems to refer to
the data that is actually sent by the hardware and not to the data used
in the HCI command, as our original implementation assumed. At least
there seem to be controllers who won't accept such "short" packets.

Change-Id: I732cfc1e05fc135fd336d832e725d969c98b60cf
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>