Android: Suppress failed connect/close attempts due to targetLost detection
Android does not provide an API to detect when an NFC tag is removed
from the vincinity. To detect such a case the Android implementation
uses active polling. During each polling cycle the implementation
connects and disconnects from the tag. If it failed, the tag is assumed
to have been removed from the vincinity.

Such failed attempts cause an IOException which are printed to stderr. This
is confusing as the developer might mistake them for some other serious
problem. This change suppresses all related warnings.

Change-Id: I95cf57076139e7d0a5ad31d4cc770a81ced12242
Reviewed-by: Frank Meerkoetter <frank.meerkoetter@basyskom.com>
Reviewed-by: Peter Rustler <peter.rustler@basyskom.com>
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>