QBluetoothServiceDiscoveryAgent: set q_ptr in d_ptr's constructor.
The agent class constructors were initializing the q_ptr member of
their d_ptr; it is cleaner to pass this down to the d_ptr's
constructor and let it do this itself.  This also lets Coverity know
that initialization actually does happen (CID 22330).  In particular,
it makes sure we can't leave it uninitialized, if an agent
implementation happens to neglect to do so.

Change-Id: Ie01046a5a113b5669e8e63c6a22f692cd3943ac0
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>