QtBluetooth (iOS) - fix potential linker errors
Bluetooth on iOS needs some of 'default' (or 'dummy') implementations,
but does not want warnings like 'Dummy backend running', as it's
not really dummy. Also, having printDummyWarning calls
can end-up in a linker error since we do not build dummy_helper.cpp on iOS.
The solution is to ifndef all these calls/includes.

Change-Id: I9c64f2bacbc6134ed30d75f85f9df0194add418d
Reviewed-by: Alex Blasche <alexander.blasche@theqtcompany.com>