Improve failed ::write() handling in QBluetoothSocket
When the Bluetooth connection is flooded with information, EAGAIN is returned.
So far this was considered an unrecoverable error which is not consistent
with the meaning of EAGAIN.

In buffered mode we'll put the data back into the buffer and in
unbuffered mode the QIODevice::write() call returns 0 which leaves the
developer the option to retry.

At the same time this patch ensures that write() calls always return
the actual number of written bytes and not the amount of bytes the caller
asked for to be written.

Partial writes in buffered mode are returned to the buffer too.

Task-number: QTBUG-54475
Change-Id: Iffc7980d2477e1fc0b45808431af6b7fecd38ded
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@theqtcompany.com>