Highlight incomplete handling of BTLE packet writes
Cases such as partial writes were silently ignored.
Since QLowEnergyController works in unbuffered mode such incomplete
writes are ignored. The ATT layer will automatically recover from such
packets on the line.

The EAGAIN case is a similar case as it effectively
means the BTLE connection is still ok but the packet can temporarily
not be written anyway. In fact, previously QBluetoothSocket reported such cases by
returning -1 as a result of QBLuetoothSocket::write(). This was even worse
as it caused a drop of the connection without justification.

These issues are reproducable when the BTLE connection is flooded
with lots of notifications.

Task-number: QTBUG-54475
Change-Id: I2b6be555ff676c440971981db0974be83c901eaa
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>