Android: Avoid newDevice signal spamming in QBluetoothDeviceDiscoveryAgent
SDP based scans and LE scans can find the saame device (address) with
different device names. This is caused by the fact that different
channels are used to retrieve the information. Those two channels
can advertise different device names.

Before this patch the first entry was added to the list of discovered
devices. When the second entry with different name came about, the
entry was never added to the list of discovered devices. Nevertheless
the deviceDiscovered() signal was triggered each time. This lead to a
situation where application which were listening to the signal only
never noticed the duplicated entries and showed a new discovered
device each time the address matched but not the remainder of the
device info.

This problem was made worse by the fact that the LE scan on Android
continues to fire the new device discovered signal once per second.

This patch distinguishes devices with the same address but different
names. They are treated as separate entries now.

Change-Id: I77259a888708309338110831e86bb7ce9253f2a6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@theqtcompany.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>