Fix NFC corkboard example on Android
The example uses a custom Manifest. The original version was based on the
general AndroidManifest.xml template at the time. In the mean time
the template changed (in particular the minimal SDK version was increased
to version 16). THe old custom manifest required API lvl 14 which was
too low.

The new version is based on the current AndroidManifest.xml template
which currently is Qt version 5.7.1.

Change-Id: I91c04cba8c7a9979d1cff0dc070e42b98185872f
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>