LE peripheral - fix a characteristic's length
With Qt's Bluetooth API it's possible to write values with different
lengths into the same characteristic. Imagine this scenario:

   service->writeCharacteristic(handle, QByteArray("abcd"));

and then

   service->writeCharacteristic(handle, QByteArray(2, 0));

If we read from the characteristic later, we get:

   00cd.

A write request in CoreBluetooth is presented as an array of CBATTRequest,
each of them has its offset and value with  length. After processing these
requests we truncate a characteristic's length if needed.

Task-number: QTBUG-56898
Change-Id: I674637eb78c806364d1a89a1db8ab3f31a8800ce
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>