Don't use "socketChannel" if the "getServiceChannel" call failed
getServiceChannel() does not exist on more recent Android platforms.
The JNI call returns an exception and socketChannel is either 0 on debug
or undefined in release builds. The previous code did not reliably
recognize the case "0" (debug) or "random int" (release) as an error
state and hence channel variable was overridden with wrong value.

Change-Id: Ia81219376661be3fbbe0c9122f707bb82defa946
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>