Add characteristic even if the read fails
If during discovery the read of a characteristic fails, then this
characteristic was not added to the service. This happens in a special
case and not always, when we try to read from a notify only
characteristic.

Now the characteristic will be added even if the read fails.

Task-number: QTBUG-58056
Change-Id: Ib802eeb66aeae92da690c296faf57331123353e6
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>