winrt: Minor adjustments to QBluetoothLocalDevice
The WinRT API does not provide functionality to obtain the state
of the devices host mode. In order to make the examples (for example
btscanner) work in winrt, the hostMode defaults to HostConnectable
and the local device is always valid on this platform.

Additionally the API only enables us to discover paired devices,
so every device that is found, has to be in "Paired" state.

Change-Id: I32359d0defdb9f8e1cc1a629d32da1214a669aff
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>