winrt: Accept malformed profile descriptor list from applications
Other backends accept a malformed profile descriptor list which only
contains a short uuid. WinRt seems to be more restrictive here and
expects a nested sequence containing a short uuid and a version per
entry.

In order not to break existing applications the WinRT backend
repairs the profile descriptor list's structure so that it is accepted
as part of IRfcommServiceProvider's list of sdp attributes.

Task-number: QTBUG-58529
Change-Id: Ic1754135bc139a732ac014728ef9341205a59e86
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>