Permit BlueZ platforms to set the address type in heartrate-game
BlueZ is the only platform that requires to set the BTLE address type.
The patch introduces a new UI button which allows the user to toggle
the address type.

The UI element is disabled on all platforms but Linux/BlueZ systems.

The number of devices returned in simulator mode is reduced to
four devices. This was done purely for esthetical reasons as the
additional address type button takes screen real estate away and
five simulator devices require more screen real estate than the
remaining space for device list. This avoid the need for scrolling.

Change-Id: I14cf26b3c821db7768b9d018c2d09441cd97679a
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Simon Hausmann <simon.hausmann@qt.io>