Fix readvertisement failure on peripheral restart on BlueZ
When the QLowEnergyController disconnects and immediately restarts
the re-advertisement, it fails when sending the OcfLeSetAdvEnable ocf
command. Subsequenly an advertisement error is thrown which stops
the QLowEnergyCOntroller instance itself.

Starting with this patch we choose to ignore failures when trying
to disable advertisement. The API cannot report the error back anyway.

Task-number: QTBUG-58941
Change-Id: Ifd6b43193c41e6cdc22d4876441bcdcbcc2ca7b6
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>