Fix crash at access to unexistent controller object
Sometimes are possible a situations when triggered event callback function
contains a pointer with a context of unexistent controller's object (f.e.
this controller can be already deleted). That causes a crash at trying to
access by this pointer.

Solution is to append a pointer of each created object (or to remove a
pointer of each destroyed object) to (from) some guarded vector of pointers.
And then to check every received context with pointers inside of this vector.
If a context does not contains in a vector, then skip any actions with this
context.

Change-Id: I94e237c41028724307f58bb09235299eae15d6b5
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>