Android: Don't miss services if the last device discovery failed
Android collects the SDP information for all devices and puts them
into an sdp cache. Only when the discovery finishes the sdp cache is
evaluated and the serviceDiscovered() signal is sent. The evaluation is
triggerd by _q_fetchUuidsTimeout(). If it is not possible
to retrieve uuids or an address for the last discovered device,
the old code never triggered _q_fetchUuidsTimeout(). As a consequence
QBluetoothServiceDiscoveryAgent appeared to never find any service.

This patch ensures that if the last SDP discovery failed we still
trigger the final UUID evaluation and subsequent serviceDiscovered()
signal emissions.

Change-Id: I8454dc3f7cf3688110e6c86af324d797351b5b33
Reviewed-by: Christian Stromme <christian.stromme@qt.io>