winrt: Ensure that QLowEnergyController::RemoteHostClosedError is emitted
As there is no Windows API to disconnect from a BTLE device, we just set
the internal state in QLowEnergyController::disconnectFromDevice. By
removing the callback registration in disconnectFromDevice we can be sure,
that every state change from connected to disconnected is caused by the
remote host closing the connection and thus implement
RemoteHostClosedError.

Change-Id: I7dcacaffce68784cb297b02df2e9f9356bf14d9d
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>