Avoid recursion during QBluetoothDeviceDiscoveryAgent::stop()
QBluetoothDeviceDiscoveryAgent (DDA) is deleted by QBluetoothServiceDiscoveryAgent (SDA)
once it receives an error or SDA is instructed to stop() the discovery process.
Currently, this triggers at least two calls to the DDA's stop() function.
In addition, if stop() generates an error the error call itself will
call back to DDA:stop(). Therefore it create an endless loop back.
This is at least true for the Android devices mentioned in the related bug report.

This patch ensures that the main logic of Android's DDA::stop() is not called
more than once. Further more SDA disconnects from DDA's signals to avoid
a potential endless loop. Any error in DDA is not of relevance to the
surrounding SDA instance anymore.

Task-number: QTBUG-60131
Change-Id: I1df16f2b0896928833aa2ced75c43d4642b4fba3
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>