CoreBluetooth - fix compilation errors for 10.13 SDK
Apple broke source code compatibility by moving 'identifier' properties from
CBPeripheral and CBCentral classes into the new CBPeer class and marking it
as "since 10.13". As a result we have compilation errors in the code
that was valid since 10.7. To workaround this issue, we introduce a
category with 'identifier' property declarations for SDK >= 10.13
and target < 10.13.

Change-Id: Id7c9d33a9c8ae7627a06b4c1930e70cede1e5b07
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>