Add a thread for QLowEnergyControllerPrivateWin32
Create a re-usable thread ('thread') and worker ('treadWorker')
in QLowEnergyControllerPrivateWin32 that have the same lifespan
as the controller object.

Add "job" support (ThreadWorkerJob) for the thread so that jobs
are scheduled and executed sequentially.

Each job should have a data struct. For writing that is
WriteCharData.

Handle writing of characteristics in:
  QLowEnergyControllerPrivateWin32::writeCharacteristic()
  QLowEnergyControllerPrivateWin32::jobFinished()
  ThreadWorker::runPendingJob()

The above jobFinished() and runPendingJob() use a `switch`
to determine the ThreadWorkerJob type.

Change-Id: I3331e0d4adc29565a88fd792f9a54833881ea694
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>