Improve robustness of tst_lowenergycontroller test
The test checks the value of various descriptors and characteristics
of the TI Sensortag. One of those values are
ClientCharacteristicConfiguration descriptors. This type of descriptor
is consistent across multiple session for the same connecting central
device. Therefore it may change and would create frequent test failures.
Even worse, the unit test itself toggles the CCC value and assuming that
a test may fail, the CC may remain with an uncertain value.

This change makes the CCC value test more generic as it checks that
the found value is within the acceptable range w/o expecting a defined
value.

Change-Id: I004a1037ee9e602807e4df5511c7d2205c92e5fe
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>