Use WIN32_FROM_HRESULT macro to decode BLE system errors
The GetLastError() function always returns the NO_ERROR code
when any of the BluetoothGATTXXX() function fails. The reason
is that the BluetoothGATTXXX() function uses the COM API, but
the GetLastError() function uses the WIN32 API.

We need to convert the HRESULT error code to the WIN32 error
code directly, using a macro such as WIN32_FROM_HRESULT.

Note: At least, it is reproduced with
BluetoothGATTSetCharacteristicValue() function on Windows 10.

Change-Id: I548e90bb16647c885b06c32397c277255fd362f4
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>
Reviewed-by: Lubomir I. Ivanov <neolit123@gmail.com>