Don't require pairing when connecting to BLE on WinRT
If we are running on a Windows version with Creators Update or newer
(>=15063) we can use new bluetooth low energy API that allows us to
do device discovery without the requirement of having paired the
device.

Instead of connecting to the device and fetching its services, we can
use UUIDs from advertisement packages to get an initial service list.

By using that approach we can get rid of the pairing process in
'connectToDevice' and replace it with 'GetGattServicesAsync' if no
services are yet available. The old behavior was kept as a fallback for
the case where the device has already been paired.

If the backend is built with an older Windows Kit or the machine
running the application does not support the new API we fall back to
the old implementation which still requires pairing.

Task-number: QTBUG-58660
Change-Id: Ice262d4d993a0119efb8b6c8fe9a21830717569f
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>