Fix QLEController::connectToDevice() on reconnect behavior
If the desired remote peripheral is already connected, the next call to
connectToDevice() does not succeed. The code path to handle this situation
already exists it was forgotten to emit the relevant connected() signal.
This patch fixes the problem.

The problem only exists on Bluez 5.42+ platforms which use the DBus
central implementation.

[ChangeLog][Platform Specific Changes][Linux] Fixed missing emission of
QLEController::connected() upon reconnect to already connected device.
This affected platforms with Bluez 5.42+ only.

Task-number: QTBUG-68911
Change-Id: Ife2f3b41c33a142d6627e433cb3b141ce9b1ff8a
Reviewed-by: Maurice Kalinowski <maurice.kalinowski@qt.io>
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>
Reviewed-by: Oliver Wolff <oliver.wolff@qt.io>