Fix crash in QLEController caused by random device disconnect
This only happens with Bluez 5.42 or later versions as the DBus based GATT
central client is used. The peripheral disconnects in the middle of a char
or descriptor write request. The disconnect triggers the cleanup of all
internal states. The write char/descriptor dbus request returns after the
disconnect and executes in the context of an already reset
QLowEnergyController instance.

The fix ensures that the call back code does not make the assumption
of operating with a connected peripheral. In fact the code had the
relevant asserts already.

[ChangeLog][QtBluetooth][PLatform Specific Changes] Fixed crash in
DBus gatt central backend (BlueZ5) caused by device disconnecting
and pending dbus call watcher returning later after disconnect.

Task-number: QTBUG-68890
Change-Id: Iad9b8a1cfc8d916d49fd9b71b2d8f03b9c90639b
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>