win32-ble: store service handles until a service is destroyed
Recently, a problem was discovered that adding event registration
on a descriptor does not work because the associated service handle
is closed right after. For events to work a service should remain
open if an event is registered for it.

Store service handles (hService) is QLowEnergyServicePrivate.
A handle is opened by
QLowEnergyControllerPrivateWin32::discoverServiceDetails()
and remains open until the device is disconnected.

This removes the need to open/close services each time a Write/Read
operation is performed on GATT characteristics / descriptors.

Move the creation of the worker thread to connectToDevice().
Quit the thread in disconnectFromDevice() and also close any
open service handles.

Change-Id: Ia634af2e4225f5c1be93b0ddd17639c2dbd70c21
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Alex Blasche <alexander.blasche@qt.io>